feat: implement Stripe webhook handler and booking confirmation verification flow

+++ src/app/api/stripe/webhook/route.ts
@@ -100,7 +100,7 @@ export async function POST(request: NextRequest) {
                     appointmentDate: slot.date,
                     appointmentTime: slot.startTime,
                     status: 'pending_confirmation',
-                    amountPaid: parseInt(amount) || 0,
+                    amountPaid: parseFloat(amount) || 0,
                     paymentStatus: 'paid',
                     stripeSessionId: session.id,
                     isMedicare: isMedicare === 'true',
